chore: rename folders to plural and update imports
