+++ server/routes/productRouter.js
@@ -1,8 +1,18 @@
 import express from 'express';
+import {
+  getProducts,
+  getProduct,
+  createProduct,
+  updateProduct,
+  deleteProduct
+} from '../controllers/ProductController.js';
 
 const router = express.Router();
 
-
-router.get('/products', () => {});
+router.get('/products', getProducts);
+router.get('/products/:id', getProduct);
+router.post('/products', createProduct);
+router.put('/products/:id', updateProduct);
+router.delete('/products/:id', deleteProduct);
 
 export default router;
